geotiff tests: consolidate IFD parser cluster (#2426)
Fold seven top-level files into unit/test_ifd.py:

- test_ifd_chain_cap.py
- test_ifd_cycle_1913.py
- test_ifd_entry_table_bounds_1672.py
- test_parallel_strip_decode_sparse_2100.py
- test_parse_all_ifds_malformed_chain_1863.py
- test_parse_ifd_bounds.py
- test_sparse_cog.py

Sections, in failure-mode order: entry-table bounds (truncated reads,
negative offsets, classic + BigTIFF), entry-value bounds (count cap,
byte cap, value-range vs EOF, pixel-tag exemptions), chain length cap
(MAX_IFDS), chain cycle detection, malformed chain offsets past EOF,
sparse blocks via the read pipeline, sparse strips through the
parallel-decode pipeline (local + HTTP COG, planar=1 + planar=2).

The cross-file duplicates between cycle and chain-cap files
(test_offset_past_eof_still_raises, test_max_ifds_still_raises,
test_short_acyclic_chain_parses) collapse onto the canonical cases in
the new file; CLUSTER_AUDIT_IFD.md maps every old file::test to its
new file::test_id.

Cluster 2 of long-tail epic #2424. Tests-only -- no source changes.
